remove authentication handling and reorganize cluster provider logic
