+++ tests/test_utils.py
@@ -26,7 +26,6 @@ from utils.constants import (
 )
 from utils.context import (
     AppContext,
-    StaticClusterProvider,
     get_cluster_connection,
 )
 from utils.index_utils import (
@@ -38,6 +37,7 @@ from utils.index_utils import (
     validate_connection_settings,
     validate_filter_params,
 )
+from utils.provider import StaticClusterProvider
 
 
 class TestIndexUtilsFunctions:
@@ -479,7 +479,7 @@ class TestContextModule:
         }
 
         with patch(
-            "utils.context.connect_to_couchbase_cluster",
+            "utils.provider.connect_to_couchbase_cluster",
             return_value=mock_cluster,
         ) as mock_connect:
             provider = StaticClusterProvider(settings=mock_settings)
@@ -500,7 +500,7 @@ class TestContextModule:
         }
 
         with patch(
-            "utils.context.connect_to_couchbase_cluster",
+            "utils.provider.connect_to_couchbase_cluster",
             return_value=mock_cluster,
         ) as mock_connect:
             provider = StaticClusterProvider(settings=mock_settings)
@@ -519,7 +519,7 @@ class TestContextModule:
         }
 
         with patch(
-            "utils.context.connect_to_couchbase_cluster",
+            "utils.provider.connect_to_couchbase_cluster",
             side_effect=Exception("Auth failed"),
         ):
             provider = StaticClusterProvider(settings=mock_settings)
@@ -539,7 +539,7 @@ class TestContextModule:
         }
 
         with patch(
-            "utils.context.connect_to_couchbase_cluster",
+            "utils.provider.connect_to_couchbase_cluster",
             return_value=mock_cluster,
         ):
             provider = StaticClusterProvider(settings=mock_settings)
